use sprintf rather than snprintf in nfai_quoted_char
snprintf doesn't exist in plain C89, and so glibc doesn't provide it
in C89 mode unless one of the relevant feature macros is provided
(e.g., _XOPEN_SOURCE=500, _BSD_SOURCE, etc; check snprintf man page
for details).

The maximum length of the output string is known and can be checked in
advance, and nfai_quoted_char is an internal function so uses of it can
also be controlled to ensure they always provide enough buffer space.