fix a clang static-analyzer warning in nfai_alloc
The report was that in nfai_alloc, a NULL page could be dereferenced
(in the p = page->data + page->at; line).  That null dereference could
only happen if the requested size (sz) is zero, because the new page
allocation and page == NULL check is only skipped if free_size >= sz,
which cannot happen unless page is non-NULL or sz is zero.

Since sz should never be zero (we have NFAI_ASSERT(sz > 0) in the
argument checking block), this report from scan-build does not actually
indicate a bug.

However, in the interest of maintaining clean scan-build output, it seems
worthwhile to fix this by moving the (page == NULL) check out of the
new page allocation branch, so that the final accesses to page are
always properly guarded (even in NDEBUG builds).