add a simple fuzz test program
The fuzz tester (the simple random-input version) has already found a
couple of bugs. The grammar based fuzzer hasn't found anything so far,
but I'm hopeful that it will be more useful with more tuning to produce
interesting inputs, and with the addition of a mode to generate from
the grammar and then make some random character changes to add errors.

Either fuzzer will be more useful with the addition of extra internal
consistency/invariant checks within the library itself.  Since the
fuzzers don't have a mechanism for checking that the actual output
of the library is *correct*, their primary purpose is just to exercise
a variety of code paths, and particularly to give coverage for "weird"
input that might not otherwise be tested.

It would be interesting to check actual tested code coverage (with gcov
or similar) when running these fuzzers to see how well they do.