add a script to build as a shared library
This is really just to document an appropriate build command
for producing a shared library, not because libnfa *should*
be built as a shared library. libnfa is designed to be plugged
into an application directly, not linked dynamically, but
who knows what people might want in the future?

Key features of the shared lib:
- builds with debug and produces a stripped version
- applies symvers to the exported symbols
- hides non-exported symbols